[CL-3581] Split shared.rb into multiple files
